Implementation of Filteration using Unscattered tracks done, and giving satisfactory results. Added the corresponding flags in CMake to switch this functionaliy ON or OFF. Depending upon the cmake flag USE_UNSCATTERED_TRACKS=ON/OFF, the filteration functionality can be turn ON/OFF. If ON then the filtered Poca Points based on unscattered tracks will be writtent to PocaPtRunHelper-Exact.txt file, otherwise it will simply dump all the PoCA points.

+++ Helpers/inc/CommonFunc.h
@@ -73,7 +73,7 @@ static void GetTracksVector(std::string fileToRead, std::vector<Track> &incoming
 	double incomingTrackP2X = 0.,incomingTrackP2Y = 0.,incomingTrackP2Z = 0.;
 	double outgoingTrackP1X = 0.,outgoingTrackP1Y = 0.,outgoingTrackP1Z = 0.;
 	double outgoingTrackP2X = 0.,outgoingTrackP2Y = 0.,outgoingTrackP2Z = 0.;
-	double scattererHitted = 0.;
+	int scattererHitted = 0;
 	double momentum = 0.;
 
 	while(!infile.eof()){
